Move bookmark handlers to handler package to avoid circular dependencies
