+++ cmd/server/main.go
@@ -261,8 +261,8 @@ func main() {
 	svr.RegisterRoute("/company/{slug}", handler.CompanyBySlugPageHandler(svr, companyRepo, jobRepo), []string{"GET"})
 
 	// bookmarks (saved jobs)
-	svr.RegisterRoute("/profile/bookmarks", bookmark.BookmarkListHandler(svr, bookmarkRepo), []string{"GET"})
-	svr.RegisterRoute("/x/bookmark", bookmark.BookmarkJobHandler(svr, bookmarkRepo, jobRepo), []string{"POST", "DELETE"})
+	svr.RegisterRoute("/profile/bookmarks", handler.BookmarkListHandler(svr, bookmarkRepo), []string{"GET"})
+	svr.RegisterRoute("/x/bookmark", handler.BookmarkJobHandler(svr, bookmarkRepo, jobRepo), []string{"POST", "DELETE"})
 
 	//
 	// auth routes
